# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Farmers Module Service with Workflow-Based Architecture for Farm Management",
        "title": "Farmers Module API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/crops/activities": {
            "get": {
                "description": "Retrieve a list of all farm activities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm-activities"
                ],
                "summary": "List farm activities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new farm activity for a crop cycle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm-activities"
                ],
                "summary": "Create a new farm activity",
                "parameters": [
                    {
                        "description": "Farm activity data",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/crops/activities/{activity_id}": {
            "get": {
                "description": "Retrieve a specific farm activity by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm-activities"
                ],
                "summary": "Get farm activity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing farm activity details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm-activities"
                ],
                "summary": "Update a farm activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity update data",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/crops/activities/{activity_id}/complete": {
            "post": {
                "description": "Mark a farm activity as completed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm-activities"
                ],
                "summary": "Complete a farm activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity completion data",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/crops/cycles": {
            "get": {
                "description": "Retrieve a list of all crop cycles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crop-cycles"
                ],
                "summary": "List crop cycles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Start a new crop cycle for a specific farm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crop-cycles"
                ],
                "summary": "Start a new crop cycle",
                "parameters": [
                    {
                        "description": "Crop cycle data",
                        "name": "cycle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/crops/cycles/{cycle_id}": {
            "get": {
                "description": "Retrieve a specific crop cycle by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crop-cycles"
                ],
                "summary": "Get crop cycle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Crop Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing crop cycle details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crop-cycles"
                ],
                "summary": "Update a crop cycle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Crop Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Crop cycle update data",
                        "name": "cycle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/crops/cycles/{cycle_id}/end": {
            "post": {
                "description": "End an active crop cycle with status and outcome",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crop-cycles"
                ],
                "summary": "End a crop cycle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Crop Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Crop cycle end data",
                        "name": "cycle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}