# internal/handlers/crop_handlers.go
package handlers

import (
	"net/http"
	"time"

	"github.com/Kisanlink/farmers-module/internal/services"
	"github.com/gin-gonic/gin"
)

// Crop Cycle Handlers (W10-W13)

// StartCycle handles W10: Start crop cycle
// @Summary Start a new crop cycle
// @Description Start a new crop cycle for a specific farm
// @Tags crop-cycles
// @Accept json
// @Produce json
// @Param cycle body object true "Crop cycle data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/cycles [post]
func StartCycle(service services.CropCycleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req struct {
			FarmID       string    `json:"farm_id" binding:"required"`
			Season       string    `json:"season" binding:"required"`
			StartDate    time.Time `json:"start_date" binding:"required"`
			PlannedCrops string    `json:"planned_crops" binding:"required"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Crop cycle started successfully",
			"data":    req,
		})
	}
}

// UpdateCycle handles W11: Update crop cycle
// @Summary Update a crop cycle
// @Description Update an existing crop cycle details
// @Tags crop-cycles
// @Accept json
// @Produce json
// @Param cycle_id path string true "Crop Cycle ID"
// @Param cycle body object true "Crop cycle update data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/cycles/{cycle_id} [put]
func UpdateCycle(service services.CropCycleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		cycleID := c.Param("cycle_id")
		var req struct {
			Season       *string    `json:"season,omitempty"`
			StartDate    *time.Time `json:"start_date,omitempty"`
			PlannedCrops *string    `json:"planned_crops,omitempty"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Crop cycle updated successfully",
			"data": gin.H{
				"cycle_id": cycleID,
			},
		})
	}
}

// EndCycle handles W12: End crop cycle
// @Summary End a crop cycle
// @Description End an active crop cycle with status and outcome
// @Tags crop-cycles
// @Accept json
// @Produce json
// @Param cycle_id path string true "Crop Cycle ID"
// @Param cycle body object true "Crop cycle end data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/cycles/{cycle_id}/end [post]
func EndCycle(service services.CropCycleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		cycleID := c.Param("cycle_id")
		var req struct {
			Status  string  `json:"status" binding:"required"`
			Outcome *string `json:"outcome,omitempty"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Crop cycle ended successfully",
			"data": gin.H{
				"cycle_id": cycleID,
				"status":   req.Status,
			},
		})
	}
}

// ListCycles handles W13: List crop cycles
// @Summary List crop cycles
// @Description Retrieve a list of all crop cycles
// @Tags crop-cycles
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/cycles [get]
func ListCycles(service services.CropCycleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Crop cycles retrieved successfully",
			"data":    []interface{}{},
		})
	}
}

// GetCropCycle handles getting crop cycle by ID
// @Summary Get crop cycle by ID
// @Description Retrieve a specific crop cycle by its ID
// @Tags crop-cycles
// @Accept json
// @Produce json
// @Param cycle_id path string true "Crop Cycle ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/cycles/{cycle_id} [get]
func GetCropCycle(service services.CropCycleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		cycleID := c.Param("cycle_id")

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Crop cycle retrieved successfully",
			"data": gin.H{
				"cycle_id": cycleID,
			},
		})
	}
}

// Farm Activity Handlers (W14-W17)

// CreateActivity handles W14: Create farm activity
// @Summary Create a new farm activity
// @Description Create a new farm activity for a crop cycle
// @Tags farm-activities
// @Accept json
// @Produce json
// @Param activity body object true "Farm activity data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/activities [post]
func CreateActivity(service services.FarmActivityService) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req struct {
			CropCycleID  string    `json:"crop_cycle_id" binding:"required"`
			ActivityType string    `json:"activity_type" binding:"required"`
			PlannedAt    time.Time `json:"planned_at" binding:"required"`
			Description  string    `json:"description"`
			Metadata     string    `json:"metadata"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm activity created successfully",
			"data": gin.H{
				"crop_cycle_id": req.CropCycleID,
				"activity_type": req.ActivityType,
				"planned_at":    req.PlannedAt,
			},
		})
	}
}

// CompleteActivity handles W15: Complete farm activity
// @Summary Complete a farm activity
// @Description Mark a farm activity as completed
// @Tags farm-activities
// @Accept json
// @Produce json
// @Param activity_id path string true "Activity ID"
// @Param activity body object true "Activity completion data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/activities/{activity_id}/complete [post]
func CompleteActivity(service services.FarmActivityService) gin.HandlerFunc {
	return func(c *gin.Context) {
		activityID := c.Param("activity_id")
		var req struct {
			CompletedAt time.Time `json:"completed_at" binding:"required"`
			Notes       *string   `json:"notes,omitempty"`
			Outcome     *string   `json:"outcome,omitempty"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm activity completed successfully",
			"data": gin.H{
				"activity_id":  activityID,
				"completed_at": req.CompletedAt,
			},
		})
	}
}

// UpdateActivity handles W16: Update farm activity
// @Summary Update a farm activity
// @Description Update an existing farm activity details
// @Tags farm-activities
// @Accept json
// @Produce json
// @Param activity_id path string true "Activity ID"
// @Param activity body object true "Activity update data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/activities/{activity_id} [put]
func UpdateActivity(service services.FarmActivityService) gin.HandlerFunc {
	return func(c *gin.Context) {
		activityID := c.Param("activity_id")
		var req struct {
			ActivityType *string    `json:"activity_type,omitempty"`
			PlannedAt    *time.Time `json:"planned_at,omitempty"`
			Metadata     *string    `json:"metadata,omitempty"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm activity updated successfully",
			"data": gin.H{
				"activity_id": activityID,
			},
		})
	}
}

// ListActivities handles W17: List farm activities
// @Summary List farm activities
// @Description Retrieve a list of all farm activities
// @Tags farm-activities
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/activities [get]
func ListActivities(service services.FarmActivityService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm activities retrieved successfully",
			"data":    []interface{}{},
		})
	}
}

// GetFarmActivity handles getting farm activity by ID
// @Summary Get farm activity by ID
// @Description Retrieve a specific farm activity by its ID
// @Tags farm-activities
// @Accept json
// @Produce json
// @Param activity_id path string true "Activity ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /crops/activities/{activity_id} [get]
func GetFarmActivity(service services.FarmActivityService) gin.HandlerFunc {
	return func(c *gin.Context) {
		activityID := c.Param("activity_id")

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm activity retrieved successfully",
			"data": gin.H{
				"activity_id": activityID,
			},
		})
	}
}